+++ docs/HISTORY.rst
@@ -4,6 +4,9 @@ There's a frood who really knows where his towel is
 1.0a8 (Unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
+- Added section method in ExternalBlogEntry (closes #7745) 
+  [marcosfromero]
+
 - Avoid error if text is empty [cleberjsantos]
 
 1.0a7 (2013-08-02)